DEFAULT_SEED: set to uint32 matching _set_global_seed

+++ psyneulink/core/components/functions/function.py
@@ -200,7 +200,7 @@ EPSILON = np.finfo(float).eps
 
 # numeric to allow modulation, invalid to identify unseeded state
 def DEFAULT_SEED():
-    res = np.array(-1)
+    res = np.array(-1).astype(np.uint32)
     print('DEFAULT SEED TYPE', res.dtype)
     return res
